Adding networktables.json to .gitignore
It seems to me that networktables.json is written by the dashboard (or robot or whatever) and is transient and is going to cause us grief if we track it. At the moment, we don’t use it for anything. I think it’s just a comms channel for the running robot.