Preliminary merge of arm into swerve code
Include a wrist start position in addition to the (now misnamed) MIN and MAX.  Need to have wrist in the start position when rebooting code.
Tested wrist and arm code.
Swerve drive modified for now to initialize the Falcon turning encoder to zero instead of using the absolute encoders.  This means that the wheels need to be aligned before enabling.  We'll need to change that to reset in autonomousInit() or go back to using the cancoders.  
The issue with CANCoder 4 seems to be that it doesn't always respond to the getAbsolutePosition() command and the Falcon's encoder gets set to ???  We see CAN wires not fully twisted, so perhaps that affects the signal quality.  More stuff to test...