Added temporary steering diagnostics
Run this and observe the output on the console.  During module init, it prints the return value of the configOffset command (an error or no error) and then the value returned by getAbsolutePosition(), which is what the Falcon's encoder would be set to.  Then at the end of RobotInit(), it asks for the absolute position 10 times for each module, spaced by 0.1 second.  Check to see if they are all the same, or if the FR has anomalies.